third_party/blink/web_tests/platform/generic/fast/dom/Element/class-list-move-between-document-with-different-quirks-mode-expected.txt

Tests that moving an element between documents with different quirks mode works

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

In standards mode
PASS htmlElement.classList.contains("A") is true
PASS htmlElement.classList.contains("a") is true
PASS htmlElement.classList.contains("B") is true
PASS htmlElement.classList.contains("b") is false
Moved to quirks mode
PASS htmlElement.classList.contains("A") is true
PASS htmlElement.classList.contains("a") is true
PASS htmlElement.classList.contains("B") is true
PASS htmlElement.classList.contains("b") is false
Moved back to standards mode
PASS htmlElement.classList.contains("A") is true
PASS htmlElement.classList.contains("a") is true
PASS htmlElement.classList.contains("B") is true
PASS htmlElement.classList.contains("b") is false
PASS successfullyParsed is true

TEST COMPLETE
